+++ src/components/mask/mask-input.js
@@ -183,6 +183,7 @@ MaskInput.prototype = {
 
     // Store an initial value on focus
     this.focusEventHandler = function () {
+      self.state.previousMaskResult = self.element.value;
       self.state.initialValue = self.element.value;
     };
     this.element.addEventListener('focus', this.focusEventHandler);
@@ -423,7 +424,7 @@ MaskInput.prototype = {
    * @returns {boolean} whether or not the previous mask state matches the current one.
    */
   _hasChangedValue() {
-    if (!this.state || !this.state.previousMaskResult) {
+    if ((!this.state || !this.state.previousMaskResult) && this.state.previousMaskResult !== '') {
       return true;
     }